requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
msal>=1.22.0
httpx>=0.24.1
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
starlette>=0.27.0
pydantic>=2.0.0
